fix: Fixes after merge
