+++ src/main/java/faang/school/accountservice/exception_handler/GlobalExceptionHandler.java
@@ -1,15 +1,20 @@
 package faang.school.accountservice.exception_handler;
 
 import faang.school.accountservice.dto.ErrorResponse;
+import faang.school.accountservice.exception.AccountNotFoundException;
+import faang.school.accountservice.exception.AccountStatusException;
 import faang.school.accountservice.exception.DataValidationException;
 import faang.school.accountservice.exception.ResourceNotFoundException;
 import faang.school.accountservice.exception.ServiceUnavailableException;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import java.time.Instant;
 
 @Slf4j
 @ControllerAdvice
@@ -31,7 +36,11 @@ public class GlobalExceptionHandler {
 
         return ResponseEntity
                 .badRequest()
-                .body(new ErrorResponse(VALIDATION_ERROR, "Validation failed", errors));
+                .body(ErrorResponse.builder()
+                        .code(VALIDATION_ERROR)
+                        .message("Validation failed")
+                        .details(errors)
+                        .build());
     }
 
     @ExceptionHandler(DataValidationException.class)
@@ -82,5 +91,35 @@ public class GlobalExceptionHandler {
                 .body(ErrorResponse.builder()
                         .code("INTERNAL_ERROR")
                         .message("An unexpected error occurred: %s".formatted(ex.getMessage())));
+    }
+
+    @ExceptionHandler(AccountNotFoundException.class)
+    public ResponseEntity<ErrorResponse> handleAccountNotFound(AccountNotFoundException e, HttpServletRequest request) {
+        log.warn("Account not found: {} | Request URI: {}", e.getMessage(), request.getRequestURI());
+        return buildErrorResponse(e, HttpStatus.NOT_FOUND, request);
+    }
+
+    @ExceptionHandler(AccountStatusException.class)
+    public ResponseEntity<ErrorResponse> handleAccountStatus(AccountStatusException e, HttpServletRequest request) {
+        log.warn("Invalid account status: {} | Request URI: {}", e.getMessage(), request.getRequestURI());
+        return buildErrorResponse(e, HttpStatus.BAD_REQUEST, request);
+    }
+
+    @ExceptionHandler(IllegalStateException.class)
+    public ResponseEntity<ErrorResponse> handleIllegalState(IllegalStateException e, HttpServletRequest request) {
+        log.error("Illegal state: {}", e.getMessage());
+        return buildErrorResponse(e, HttpStatus.BAD_REQUEST, request);
+    }
+
+    private ResponseEntity<ErrorResponse> buildErrorResponse(Exception e, HttpStatus status,
+            HttpServletRequest request) {
+        ErrorResponse errorResponse = ErrorResponse.builder()
+                .timestamp(Instant.now())
+                .status(status.value())
+                .code(status.getReasonPhrase())
+                .error(e.getMessage())
+                .path(request.getRequestURI())
+                .build();
+        return new ResponseEntity<>(errorResponse, status);
     }
 }
